libhttpseverywhere: Fix for Linuxbrew
Closes #13825.

+++ Formula/libhttpseverywhere.rb
@@ -3,13 +3,13 @@ class Libhttpseverywhere < Formula
   homepage "https://github.com/gnome/libhttpseverywhere"
   url "https://download.gnome.org/sources/libhttpseverywhere/0.8/libhttpseverywhere-0.8.3.tar.xz"
   sha256 "1c006f5633842a2b131c1cf644ab929556fc27968a60da55c00955bd4934b6ca"
-  revision 2
+  revision OS.mac? ? 2 : 3
 
   bottle do
-    cellar :any
-    sha256 "7ed8d7cc4934185649bb1d3efc00f0ffac955310085db2cf44434cdf476a7e8d" => :mojave
-    sha256 "109e95eba0ebfe5a6d3c358055bc07861e2dce33e2376aa33843e73fba9a52eb" => :high_sierra
-    sha256 "f161a27096199a9810416b47d93c1bbf95b3948641ab009a7cfb049e5264a76f" => :sierra
+    sha256 "332c2aad0bcc26b7e29c8fb77c4b4a3dc842d4e769aa323b4a39e298106e2516" => :mojave
+    sha256 "2ce35886b26f24cf347feb2a603f75c26ed5a3ab81853e5e42a7494dbe0993b8" => :high_sierra
+    sha256 "6b532d6d32e37b39e3ceeacb2914d23a8ef1b59347203a7c499cb39fb320a8ea" => :sierra
+    sha256 "c6dc35ab2114924942f08fd412b4b5e0ea34751c6d72e3642a2b26614b8dc3b3" => :el_capitan
   end
 
   depends_on "gobject-introspection" => :build
@@ -26,8 +26,8 @@ class Libhttpseverywhere < Formula
   # see https://gitlab.gnome.org/GNOME/libhttpseverywhere/issues/1
   # remove when next version is released
   patch do
-    url "https://gitlab.gnome.org/GNOME/libhttpseverywhere/commit/6da08ef1ade9ea267cecf14dd5cb2c3e6e5e50cb.patch"
-    sha256 "eeb4f672c6cee3db6f7ba0f0798e37be68f1d3d6d8e22fb0ec21e7ca768356c3"
+    url "https://gitlab.gnome.org/GNOME/libhttpseverywhere/commit/6da08ef1ade9ea267cecf14dd5cb2c3e6e5e50cb.diff"
+    sha256 "e5499c290c5b48b243f67763a2c710acc5bd52b90541eb8da3f8b24b516f7430"
   end
 
   def install
@@ -37,12 +37,14 @@ class Libhttpseverywhere < Formula
       system "ninja", "install"
     end
 
-    dir = [Pathname.new("#{lib}64"), lib/"x86_64-linux-gnu"].find(&:directory?)
-    unless dir.nil?
-      mkdir_p lib
-      system "/bin/mv", *Dir[dir/"*"], lib
-      rmdir dir
-      inreplace Dir[lib/"pkgconfig/*.pc"], %r{lib64|lib/x86_64-linux-gnu}, "lib"
+    if OS.mac?
+      dir = [Pathname.new("#{lib}64"), lib/"x86_64-linux-gnu"].find(&:directory?)
+      unless dir.nil?
+        mkdir_p lib
+        system "/bin/mv", *Dir[dir/"*"], lib
+        rmdir dir
+        inreplace Dir[lib/"pkgconfig/*.pc"], %r{lib64|lib/x86_64-linux-gnu}, "lib"
+      end
     end
   end
